feature (app): create and save a value
